fail: does not work
this was an attempt to manually avoid some reference counting invocations by fiddling with the function calling path through eval. it *might* work with some more effort, but for now i'm abandoning it. it seems to get progressively more baroque as i go along, making fast, simple things not so fast and simple any more.