cruft removal
attempt to fix bug, i think it was possible for two background processes to get going at once which might result in the double-send bug i was seeing. also, on older androids, it is possible that "external storage" permission is granted but external storage might not be accessible. in which case, we fall back to using the internal storage. this isn't the best, but at least the app won't crash in these cases.