wip, cruft removal
removal old protocols. note: it might make sense to limit the number of retries over a period of time. if a file is corrupted for example, clients might try to retry and we would keep the corrupt bits as an "automatic restart". maybe after a certain number of retries result in no progress, just unlink the file so the transfer happens again.