turn off sorting attempt
this type of sorting might be useful for something else, so i'm leaving it in for now. but it was too complicated to get right. a better approach might be to just either filter out messages, or provide a button that pulses to go to the next image that has new info.