BUG in qtcreator
previously, the targetSdk was set from the androidmanifest, without specifying it anywhere else (like qmake file). now, it ignores the androidmanifest, and ends up using the default compiled into qmake (which is 28, despite what the docs say.) you need to specify the ANDROID_TARGET_SDK_VERSION explicitly in the qmake file to get the gradle.properties file to have the right thing in it.