wip
if there is an error on sync channel, just nuke the entire control channel and whatnot, we won't be able to recover piecemeal. added ping check too, not sure if this will be useful, but was having problems with nat-nat action where some of the channels would lock up for some reason.