wip, api changes
try to remove from the input api all the instances of "uid" where it wasn't a filter, but rather, just an indication of where some mid was stored. since we are more or less wed to the idea of using the message index as an integral part of message processing (and not just a side thing), this should be ok.