tweaks for character encodings on old messages
note that this will only work for messages and text (in profiles and so on) that originated from the ancient times before unicode/utf8. since the messages and text stored in the database were all from windows vintages that used code pages similar to "latin1", this helps in displaying those old things without a lot of dropped or mis-displayed characters. the real fix would be to check to see if a utf8 conversion resulted in a broken character, and apply latin1 instead (and maybe attempt some kind of age determination as another heuristic.) i think most of the input to the system now is in the form of utf8, but it isn't "enforced", so i really need to find all the spots where data is originated in unicode and convert explicitly to utf8 or something.