make unviewed msg handling more sensible
if the server returns old messages from before you are in the group, but others in the group have seen the msg, this just filters that out. really the server should try harder not to return old messages. it should probably notice when you enter the group, and enter enough tombstones on your behalf for the existing messages it has.