clear all pull asserts when changing eager mode
if we don't clear them, the next connect will clear the failures and try to reacquire lots of messages. which begs the question, should we just put a timer on the failures instead of clearing them all? may not be worth the complexity. the set of asserts for eager/lazy pulls is completely different so clearing them on mode change makes sense.