we were getting two incoming syncs from same uid
this one is hard to test since it requires just the right combo of network failures and ip reassignments. seems like if you switch one back and forth out of a vpn it might cause one of the parties to close and retry, while the remote party doesn't get notification the connection is a zombie.