more android migrate
jesus fuck me with a brick. we have to target 30, and add "preserveLegacyExternalStorage" to be able to READ our previous data. this only works if the app hasn't been uninstalled. so, if you uninstall, you lose everything. i'm not overly worried about this, if i get an irate user, i can always manually send them the pics, since they are all stored on the server.