unify dir layout a little
phoo used a "tmp" dir, but cdc-x did not. dwycobg would get confused if the tmp dir didn't exist while doing background ops for cdc-x. now everyone uses the same layout. had to fix bug in set_fn_prefixes. also, exposed a bug in the sync protocol setup where it failed to create a tmp database, causing the protocol to get stuck in "recv_init" waiting for a delta that never got sent. this needs to be fixed.