horrible hack for windows
at least it makes it plausible to launch an full innosetup installer from qdwyrun. none of this would be a huge problem if qdwyrun wasn't dependent on qt, or i could recompile the waitrun stuff. anyway, this is ok for now. note: the simple update stuff will still work if all you need is to patch a few exe's or whatever.