might be a fix for message tombstones
when we see a new tombstone, cancel pulls and ongoing operations to pull that mid from remotes. this might fix some of the weird things we see when deletes get stranded on some clients (though, it was still acting a little weird during testing.) this fixes a problem where we might try to pull something, then a tombstone shows up later from another client.