try perodic work instead of one time work
this seems to work ok, but may require some tuning. the android background tries to sense when no messaging is occurring and quit after a certain amount of time. but, this basically makes the client ineligible to receive sync connections in the background on demand. may want to monitor the notifications from google and restart the work at that point.