fix problem with compatibility with older style single-pk
some parts of the system use the old single-pk store and forward functions, and this was causing some problems logging in/group creation challenges. probably need to migrate away from that old format, but that might require a few releases. PROBLEMS! if you release code with this fix in it, you have to upgrade the interpreter too! what a mess. really need to thing about it a little harder before releasing anything or upgrading any of the server stuff.