qt 6 port
important things to remember:
qt6 is 64bit only, so the old video capture driver dll (32bit) will never work. if the qt6 driver cannot be made to work, this is a dead end. so, do that first.
the theora assembly code is x86. aarch must be disabled on mac until we can update makefiles.