fix http download
our new http server does a bunch of redirects to get https going by default.  this makes it follow the redirects automatically. in this case, we really need to avoid doing the https redirects since some clients may not have the capability of doing https, and this is only used in dire situations.