fix audio call issue with mac
i guess on the mac when you perform a "start" while the output is in idle state, it actually changes the state to "stopped" first. not clear from the docs if this is true in general, but other platforms seem not to have this problem. also, "idle" state seems to be similar to "active". so, the hopeful fix is to only "start" when the stream ends up in the stopped state for some reason.