wip
i started this in order to fix some problems where you could do some operations on a uid in your conversation list, and the user would disappear, and make it hard to find them again. i thought it would be quick change, but it got out of hand. it needs a rethink.