add a typing indicator near the text input
on mobile, you can't always see the top of the box where the current indicator is. this is a kluge for now.
i'm including the BlockLoader, but it is completely broken for anything but rows = 2 and columns = 4, i think it has something to do with the timing of how the repeater elements are processed after the first cycle of the animation. typical qml monkeybiz