wip, android
the permissions stuff is commented out just to get it to link. unfortunately, a good bit of the android build and deploy stuff will have to be tweaked or redone to get it working with qt 6.5. sadly, the state of qt6.5 on android doesn't inspire confidence: many of their multimedia examples will not work (even with r25b, they definitel do NOT work with 21.3xxx, getting the emutlsxxx undefined link issue). couple this with the lack of permissions for POST_NOTIFICATION (and their new framework doesn't really allow one to extend it, as far as i can see), and android update to 6.5 is kinda a waste.  updates to 6.5 for desktop seem closer though.