use macos as host for android build
ndk-which is broken in ndk 25.1 on macos. it doesn't appear to be broken on linux hosts. but, it was deprecated some time ago when gcc was deprecated. unfortunately, i get the same goofy app startup crash using macos. so, whatever is trying to load a "qt5" method and crashing in qtnative is probably not an sdk/ndk misconfig.