see if my old codec still works
sorta: it has no problems decoding old videos, but the coder is broken (ie seems to product some incorrect "motion_comp" control stream.) there are so many things wrong with it, that it can only be described as "utterly broken". fixed sizes for frames that are small, limits on RLC lengths, no checking on buffer sizes at all, mostly ripped out quantization policy stuff, ripped out "codec tweaker" setup. even if all that were fixed, the "multi-codec selection" stuff in the core is utterly broken too... ie, selecting the codec seems to be done when preview is a created rather than stream start, which means you would have to move to a new CHAT server with the other codec configured, and start and stop video capture (duh). there are references to "grayscale" codecs that i'm sure don't exist any more, etc.etc. which basically means, single-codec operation with theora is more or less wired in.

this little side branch was just a feeble attempt at debugging it a little bit and seeing if i could get it working. but it isn't worth it. it is just too crufty, and the encoder should be taken out completely.