sunset the dwyco DCT based video encoder
leaves the video DEcoder intact so old messages can still be played. we haven't actually generated any new video using this coder for over a decade. an attempt to resurrect it confirmed it was utterly broken (can't really produce a valid stream any more) so there is no point in having it around.  it was ok for its time, but it's  adios amigo.