bug hunting, stuck in-progress pulls
when the connection is closed, the sendq for that connection may have commands that refer to the state of some asserted pulls. make sure we clean up and set the "inprogress" to false when the sendq is destroyed on connection drop. note in general, anything in that queue probably needs to be checked for whatever state it might have elsewhere. right now, i can't think of anything, but possibly it needs a facility for firing some kind of "cleanup" for any commands left in the queue. alternately, it might make sense to get rid of the "in progress" state on the pulls. i'm not sure if it is needed for anything but avoiding redundant queuing.