quick hack for announcement bot
you can put a chat lobby index as the last command line arg. this is useful since there are a lot of user that just hang out in lobbies other than the default, and would never get an announcement using the old greetbot. still has to be done by hand, so useful for more urgent announcements.