turn some errors into recoverable soft errors
usually, if we got an error like trying to compare a string to a set, we terminate the program (in LH, it quits with an error. if LH not running, the lib would call panic, essentially terminating the program.) this is a problem if we are ingesting strings that might cause these "soft errors" to occur. this just makes the deserialization error out with a parse error instead of terminating the program.