cmake
builds on macos, need to port up camera permission though. WARNING: the resource stuff is really fucked up the cmake build. the docs are not clear about what all prefixes and other stuff are put on the resource paths by default. like sometimes it puts /qt/qml/uri_yada which of course borks any mention of the paths in the source code.