fix image centering for 6.7
for some reason, specifying the x coord and the width explicitly for the image delegate stopped working in 6.7. it worked fine in qt5 and qt 6.5. in qt 6.7, a model reset would cause the x coord to become 0, and on repopulating the model would not recalculate the x coordinate, which seems like a bug. it is like it is reusing a delegate without properly causing properties to reset/change. the "fix" is  probably wrong too, since i've run into problems where the "parent" of a delegate isn't defined when you expect it to be... sigh.