redo a couple of dialogs to avoid pulling in all the goofy style plugins
oddly, the qmlimportscanner scans *all* subdirectories too, i guess it figures it is better to pull in more than less. but, if you have an "old" subdirectory, it scans that too and pulls in the garbage that is referenced there. i confirmed this by chmod 0 on the dirs i didn't want it to scan, and it worked. unfortunately, even if you reference just "Material" style, it still loads the "Basic" plugin, i guess as a fallback.