fix problem with deserialization
i had a note about this in my design for this thing, but forgot to deal with it, since the only deployed thing is single-threaded, blocking app.

but, even with the limits set to *huge* (ie, don't do any checking) there were problems with threads stepping on each other. since we have decided that a single deserialization attempt is a one-off operation that would never span threads, we can just use a thread_local to isolate things. it *appears* to have fixed things (which only showed up on macos for some reason.)