weird backup bug in android
turn off backups, there is a bug that manifests itself as causing the app to hang on the splash screen after a phone backup, IF the app was in the background and android had to kill the app to perform the backup. this happens on emulators and phones. it seems to affect flutter apps too (https://stackoverflow.com/questions/71198034/flutter-android-app-freezes-on-the-splash-screen-in-release-mode).