fix priority of pull_response
this was getting the wrong index for the priority we wanted to the "response" for a pull. it was crashing because of non-const indexing bug, but before that was a problem, it would index and get "nil", which made the priority 0. which means it could get put into the queue in an odd place. it would work, but might muck up initialization since 0 priority is PULLPRI_INIT using during connection setup.