adjust permissions for reading images
for api 29 and below, we request READ_EXTERNAL_STORAGE. if >29, we have removed the request for the "READ_MEDIA_IMAGES" permission (as it is always denied.) it *appears* that android will pop up a picker without us changing anything. which makes me wonder if i really need to include the updates to the build.gradle to include the androidx.activity stuff. since i don't use it directly, this must be a compat thing google has implemented? fiik. also, api 29 is schizo, it still has the "storage permission" stuff. but, the picker it pops up doesn't produce content that is accessible unless you ask for the old permission, unlike api 30+, where the picker just defaults to "you can access the picture that is picked, and nothing else." dunno.