add some more debugging to DwSVec
this just helps file situations where you might end up with a dangling reference. this class wasn't really intended for use outside a specific case i had in "vc", but it made it out into the wild anyways. really need to get rid of the "get a reference to the internals" operations.