assortment of minor performance things
the only sketchy one is creating an explicit index when importing a new message index. that is sort of a "schema change" that might not be worth the troubles. this is a good start, but it is just time consuming to merge the data from a remote client. this probably needs some clever re-thinking.