allow broadcast even when not in a group
not sure what i was thinking here. selfs and cdcx could both use direct messaging on local network if this worked. note: there are a lot of things that need to be checked regarding totally local stuff. like if you can't get to a server to create an account, what do we do. etc.