wip, pk verification
i think i've decided that when the pk is stored, we do the verification and store the result. when the pk is later used, it gets loaded into a cache. this is where the record should be checked. you can also check manually at use-time. signals are emitted right now, but there is nothing receiving them at this point.