revert salting, allow pk if we have keys.
for now, i decided that salting doesn't meaningfully improve on weak passwords. instead, if we have pk's use them (before, we inhibited it in the interest of allowing the protocol to complete more often.) now, we check if we have a group key for the target group. if we do, great, we can send the message with pk. if not, we revert back to just using the password for encryption. note, if we wanted to improve the chances of having the group key, we might need to put some kind of "try to fetch it first" state in front of start_gj.