get the test thread working again
turns out, gst-launch pukes from dumb messages into the stream before outputting the video data. use -q to avoid that. also, the nv12 stride handling now assumes the stride from the luma plane and chroma planes are the same. though it is a little weird, since the two chroma plans are packed, so the stride applies to the entire packed chroma area.